feat: 질문 글 목록 보기 구현 (#4)

+++ RevUp-domain/src/main/java/com/revup/question/entity/Question.java
@@ -29,7 +29,7 @@ public class Question extends BaseTimeEntity {
     @Column(columnDefinition = "TEXT")
     private String content;
 
-    private int readCount;
+    private Long readCount;
 
     @Enumerated(EnumType.STRING)
     private BooleanStatus isAnonymous;
@@ -50,7 +50,7 @@ public class Question extends BaseTimeEntity {
         this.type = type;
         this.state = state;
         this.content = content;
-        this.readCount = 0;
+        this.readCount = 0L;
         this.isAnonymous = isAnonymous;
         this.user = user;
     }
